+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # 1.5.x
 
+## 1.5.3
+- since
+- Routing Middleware:
+	- Bereiche mit Methods aufgeteilt, damit die Async Mw nicht alles doppelt machen muss
+- Dispatcher: 
+	- bei 405 bzw. 404 werden doppelte Http Methods nur noch einmal durchsucht
+
 ## 1.5.2
 - since 2020/01/13
 - CallableResolver aufgeteilt in ClosureResolver und CallableResolver